Added willUpdateDataModel method to TLIndexPathControllerDelegate
The willUpdateDataModel delegate method provides an opportunity for the receiver to supply an alternate data model before the updates are applied. This can be useful for intercepting internally generated data models when the controller is configured with an NSFetchRequest and making modifications like mixing in additional (possibly non-Core Data) items. Another example is displaying a special "No results" message when there are no items to display. See the "No Results" sample project, which displays a no results message using the TLNoResultsTableDataModel extension.