Detect updated core data items in TLIndexPathUpdates (fix issue #19)
When using TLIndexPathController with Core Data, updated managed objects were not being detected because the old and updated data models provided to TLIndexPathUpdates contained the same instance. With this fix, TLIndexPathController listens to the updated item notifications coming from the internal NSFetchedResultsController and passes the list of updated items to TLIndexPathUpdates.