Change TLIndexPathController to always set new data model
Previous behavior of setDataModel was to only set the new model if isEqual was false. However, this assumes that data model items have a proper isEqual implementation, which is often not the case. The new behavior is to always set the new data model regarless of whether any changes are detected.

Note that if data model items don't have a proper isEqual implementation, then modified items won't automatically trigger batch updates. An easy workaround for this in TLTableViewController and TLCollectionViewController is to call reconfigureVisibleCells in didUpdateDataModel.