Fix nullability annotation on block-based initializer
Both blocks in the block-based initializer should be able to return nil. For section names, items are not required to have a section name. For identifiers, there are cases where the block may be called on items that aren't contained in the data model and may be of an unexpected type, in which case it is appropriate to return nil.